Refactor 13 - utility package
As multiple utility classes have been created, its better for classification purposes to group them together, thus the creation of this package. Within includes :
- Collisions
- KeyHandler
- PlaneFactory
- ProjectileFactory
- ProjectileProduction

classes affected by these changes have also been modified accordingly
