+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/levels/LevelOne.java
@@ -1,10 +1,14 @@
 package dev.ccr.dmscw2024.levels;
 
 import dev.ccr.dmscw2024.fundamentals.ActiveActorDestructible;
-import dev.ccr.dmscw2024.planes.PlaneFactory;
+import dev.ccr.dmscw2024.utility.ActorManager;
+import dev.ccr.dmscw2024.utility.KeyHandler;
+import dev.ccr.dmscw2024.utility.LevelInitialiser;
+import dev.ccr.dmscw2024.utility.PlaneFactory;
 import dev.ccr.dmscw2024.planes.enemies.EnemyPlane;
+import javafx.scene.Group;
 
-public class LevelOne extends LevelParent {
+public abstract class LevelOne extends LevelParent {
 	
 	private static final String BACKGROUND_IMAGE_NAME = "/dev/ccr/dmscw2024/images/background1.jpg";
 	private static final String NEXT_LEVEL = "dev.ccr.dmscw2024.levels.LevelSW";
@@ -12,50 +16,6 @@ public class LevelOne extends LevelParent {
 	private static final int KILLS_TO_ADVANCE = 10;
 	private static final double ENEMY_SPAWN_PROBABILITY = .20;
 	private static final int PLAYER_INITIAL_HEALTH = 5;
-
-	/*@Override
-	protected LevelParent instantiateLevel(double screenHeight, double screenWidth) {
-		return new LevelParent(
-				BACKGROUND_IMAGE_NAME,
-				screenHeight,
-				screenWidth,
-				() -> PlaneFactory.createDefaultUserPlane(screenHeight)
-		);
-	}
-
-	@Override
-	protected void initializeFriendlyUnits(LevelParent level) {
-		System.out.println("Initialising UserPlane for LevelOne");
-		System.out.println("getUser(): " + level.getUser());
-		level.getRoot().getChildren().add(level.getUser());
-	}
-
-	@Override
-	protected void spawnEnemyUnits(LevelParent level) {
-		int currentNumberOfEnemies = level.getCurrentNumberOfEnemies();
-		for (int i = 0; i < TOTAL_ENEMIES - currentNumberOfEnemies; i++) {
-			if (Math.random() < ENEMY_SPAWN_PROBABILITY) {
-				double newEnemyInitialYPosition = Math.random() * level.getEnemyMaximumYPosition();
-				ActiveActorDestructible newEnemy = new EnemyPlane(level.getScreenWidth(), newEnemyInitialYPosition);
-				level.addEnemyUnit(newEnemy);
-			}
-		}
-	}
-
-	@Override
-	protected LevelView instantiateLevelView(LevelParent level) {
-		return new LevelView(level.getRoot(), PLAYER_INITIAL_HEALTH);
-	}
-
-	@Override
-	public void checkIfGameOver(LevelParent level) {
-		if (level.userIsDestroyed()) {
-			level.loseGame();
-		}
-		else if (level.getUser().getNumberOfKills() >= KILLS_TO_ADVANCE) {}
-			level.goToNextLevel(NEXT_LEVEL);
-	}
-*/
 
 	public LevelOne(double screenHeight, double screenWidth) {
 		super(
@@ -67,18 +27,13 @@ public class LevelOne extends LevelParent {
 	}
 
 	@Override
-	protected void checkIfGameOver() {
-		if (userIsDestroyed()) {
-			loseGame();
-		}
-		else if (userHasReachedKillTarget())
-			goToNextLevel(NEXT_LEVEL);
+	protected LevelView instantiateLevelView() {
+		return new LevelView(getRoot(), PLAYER_INITIAL_HEALTH);
 	}
 
 	@Override
-	protected void initializeFriendlyUnits() {
+	protected void initialiseFriendlyUnits() {
 		System.out.println("Initialising UserPlane for LevelOne");
-		System.out.println("getUser(): " + getUser());
 		getRoot().getChildren().add(getUser());
 	}
 
@@ -89,20 +44,21 @@ public class LevelOne extends LevelParent {
 			if (Math.random() < ENEMY_SPAWN_PROBABILITY) {
 				double newEnemyInitialYPosition = Math.random() * getEnemyMaximumYPosition();
 				ActiveActorDestructible newEnemy = new EnemyPlane(getScreenWidth(), newEnemyInitialYPosition);
-				addEnemyUnit(newEnemy);
+				addEnemyUnits(newEnemy);
 			}
 		}
 	}
 
 	@Override
-	protected LevelView instantiateLevelView() {
-		return new LevelView(getRoot(), PLAYER_INITIAL_HEALTH);
+    public void checkIfGameOver() {
+		if (userIsDestroyed()) {
+			loseGame();
+		}
+		else if (userHasReachedKillTarget())
+			goToNextLevel(NEXT_LEVEL);
 	}
 
 	private boolean userHasReachedKillTarget() {
 		return getUser().getNumberOfKills() >= KILLS_TO_ADVANCE;
 	}
-
-
-
 }
