+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/planes/bosses/ISD.java
@@ -1,7 +1,7 @@
 package dev.ccr.dmscw2024.planes.bosses;
 
 import dev.ccr.dmscw2024.fundamentals.ActiveActorDestructible;
-import dev.ccr.dmscw2024.projectile.ProjectileFactory;
+import dev.ccr.dmscw2024.utility.ProjectileFactory;
 import dev.ccr.dmscw2024.specials.shield.ISDShield;
 
 public class ISD extends BaseBoss {
